refactor(wsLink): Follow linter guidelines
